requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0

# Use S3
ask-sdk-s3-persistence-adapter==1.0.0

# Use dynamoDB
ask-sdk-dynamodb-persistence-adapter==1.15.0

# For LINE
requests==2.24.0